// src/graphql/thegraph/__generated__/types-and-hooks.ts
export type Maybe<T> = T;
// ...
export type Bundle = {
  __typename?: "Bundle";
  ethPriceUSD: Scalars["BigDecimal"];
  id: Scalars["ID"];
};
// ...
export type Burn = {
  __typename?: "Burn";
  amount: Scalars["BigInt"];
  amount0: Scalars["BigDecimal"];
  amount1: Scalars["BigDecimal"];
  amountUSD?: Maybe<Scalars["BigDecimal"]>;
  id: Scalars["ID"];
  logIndex?: Maybe<Scalars["BigInt"]>;
  origin: Scalars["Bytes"];
  owner?: Maybe<Scalars["Bytes"]>;
  pool: Pool;
  tickLower: Scalars["BigInt"];
  tickUpper: Scalars["BigInt"];
  timestamp: Scalars["BigInt"];
  token0: Token;
  token1: Token;
  transaction: Transaction;
};
// ...
export type Collect = {
  __typename?: "Collect";
  amount0: Scalars["BigDecimal"];
  amount1: Scalars["BigDecimal"];
  amountUSD?: Maybe<Scalars["BigDecimal"]>;
  id: Scalars["ID"];
  logIndex?: Maybe<Scalars["BigInt"]>;
  owner?: Maybe<Scalars["Bytes"]>;
  pool: Pool;
  tickLower: Scalars["BigInt"];
  tickUpper: Scalars["BigInt"];
  timestamp: Scalars["BigInt"];
  transaction: Transaction;
};
// ...
export type Factory = {
  __typename?: "Factory";
  id: Scalars["ID"];
  owner: Scalars["ID"];
  poolCount: Scalars["BigInt"];
  totalFeesETH: Scalars["BigDecimal"];
  totalFeesUSD: Scalars["BigDecimal"];
  totalValueLockedETH: Scalars["BigDecimal"];
  totalValueLockedETHUntracked: Scalars["BigDecimal"];
  totalValueLockedUSD: Scalars["BigDecimal"];
  totalValueLockedUSDUntracked: Scalars["BigDecimal"];
  totalVolumeETH: Scalars["BigDecimal"];
  totalVolumeUSD: Scalars["BigDecimal"];
  txCount: Scalars["BigInt"];
  untrackedVolumeUSD: Scalars["BigDecimal"];
};
// ...
export type Flash = {
  __typename?: "Flash";
  amount0: Scalars["BigDecimal"];
  amount0Paid: Scalars["BigDecimal"];
  amount1: Scalars["BigDecimal"];
  amount1Paid: Scalars["BigDecimal"];
  amountUSD: Scalars["BigDecimal"];
  id: Scalars["ID"];
  logIndex?: Maybe<Scalars["BigInt"]>;
  pool: Pool;
  recipient: Scalars["Bytes"];
  sender: Scalars["Bytes"];
  timestamp: Scalars["BigInt"];
  transaction: Transaction;
};
// ...
export type Mint = {
  __typename?: "Mint";
  amount: Scalars["BigInt"];
  amount0: Scalars["BigDecimal"];
  amount1: Scalars["BigDecimal"];
  amountUSD?: Maybe<Scalars["BigDecimal"]>;
  id: Scalars["ID"];
  logIndex?: Maybe<Scalars["BigInt"]>;
  origin: Scalars["Bytes"];
  owner: Scalars["Bytes"];
  pool: Pool;
  sender?: Maybe<Scalars["Bytes"]>;
  tickLower: Scalars["BigInt"];
  tickUpper: Scalars["BigInt"];
  timestamp: Scalars["BigInt"];
  token0: Token;
  token1: Token;
  transaction: Transaction;
};
// ...
export type Pool = {
  __typename?: "Pool";
  burns: Array<Burn>;
  collectedFeesToken0: Scalars["BigDecimal"];
  collectedFeesToken1: Scalars["BigDecimal"];
  collectedFeesUSD: Scalars["BigDecimal"];
  collects: Array<Collect>;
  createdAtBlockNumber: Scalars["BigInt"];
  createdAtTimestamp: Scalars["BigInt"];
  feeGrowthGlobal0X128: Scalars["BigInt"];
  feeGrowthGlobal1X128: Scalars["BigInt"];
  feeTier: Scalars["BigInt"];
  feesUSD: Scalars["BigDecimal"];
  id: Scalars["ID"];
  liquidity: Scalars["BigInt"];
  liquidityProviderCount: Scalars["BigInt"];
  mints: Array<Mint>;
  observationIndex: Scalars["BigInt"];
  poolDayData: Array<PoolDayData>;
  poolHourData: Array<PoolHourData>;
  sqrtPrice: Scalars["BigInt"];
  swaps: Array<Swap>;
  tick?: Maybe<Scalars["BigInt"]>;
  ticks: Array<Tick>;
  token0: Token;
  token0Price: Scalars["BigDecimal"];
  token1: Token;
  token1Price: Scalars["BigDecimal"];
  totalValueLockedETH: Scalars["BigDecimal"];
  totalValueLockedToken0: Scalars["BigDecimal"];
  totalValueLockedToken1: Scalars["BigDecimal"];
  totalValueLockedUSD: Scalars["BigDecimal"];
  totalValueLockedUSDUntracked: Scalars["BigDecimal"];
  txCount: Scalars["BigInt"];
  untrackedVolumeUSD: Scalars["BigDecimal"];
  volumeToken0: Scalars["BigDecimal"];
  volumeToken1: Scalars["BigDecimal"];
  volumeUSD: Scalars["BigDecimal"];
};
// ...
export type PoolDayData = {
  __typename?: "PoolDayData";
  close: Scalars["BigDecimal"];
  date: Scalars["Int"];
  feeGrowthGlobal0X128: Scalars["BigInt"];
  feeGrowthGlobal1X128: Scalars["BigInt"];
  feesUSD: Scalars["BigDecimal"];
  high: Scalars["BigDecimal"];
  id: Scalars["ID"];
  liquidity: Scalars["BigInt"];
  low: Scalars["BigDecimal"];
  open: Scalars["BigDecimal"];
  pool: Pool;
  sqrtPrice: Scalars["BigInt"];
  tick?: Maybe<Scalars["BigInt"]>;
  token0Price: Scalars["BigDecimal"];
  token1Price: Scalars["BigDecimal"];
  tvlUSD: Scalars["BigDecimal"];
  txCount: Scalars["BigInt"];
  volumeToken0: Scalars["BigDecimal"];
  volumeToken1: Scalars["BigDecimal"];
  volumeUSD: Scalars["BigDecimal"];
};
// ...
export type PoolHourData = {
  __typename?: "PoolHourData";
  close: Scalars["BigDecimal"];
  feeGrowthGlobal0X128: Scalars["BigInt"];
  feeGrowthGlobal1X128: Scalars["BigInt"];
  feesUSD: Scalars["BigDecimal"];
  high: Scalars["BigDecimal"];
  id: Scalars["ID"];
  liquidity: Scalars["BigInt"];
  low: Scalars["BigDecimal"];
  open: Scalars["BigDecimal"];
  periodStartUnix: Scalars["Int"];
  pool: Pool;
  sqrtPrice: Scalars["BigInt"];
  tick?: Maybe<Scalars["BigInt"]>;
  token0Price: Scalars["BigDecimal"];
  token1Price: Scalars["BigDecimal"];
  tvlUSD: Scalars["BigDecimal"];
  txCount: Scalars["BigInt"];
  volumeToken0: Scalars["BigDecimal"];
  volumeToken1: Scalars["BigDecimal"];
  volumeUSD: Scalars["BigDecimal"];
};
// ...
export type Position = {
  __typename?: "Position";
  collectedFeesToken0: Scalars["BigDecimal"];
  collectedFeesToken1: Scalars["BigDecimal"];
  depositedToken0: Scalars["BigDecimal"];
  depositedToken1: Scalars["BigDecimal"];
  feeGrowthInside0LastX128: Scalars["BigInt"];
  feeGrowthInside1LastX128: Scalars["BigInt"];
  id: Scalars["ID"];
  liquidity: Scalars["BigInt"];
  owner: Scalars["Bytes"];
  pool: Pool;
  tickLower: Tick;
  tickUpper: Tick;
  token0: Token;
  token1: Token;
  transaction: Transaction;
  withdrawnToken0: Scalars["BigDecimal"];
  withdrawnToken1: Scalars["BigDecimal"];
};

export type PositionSnapshot = {
  __typename?: "PositionSnapshot";
  blockNumber: Scalars["BigInt"];
  collectedFeesToken0: Scalars["BigDecimal"];
  collectedFeesToken1: Scalars["BigDecimal"];
  depositedToken0: Scalars["BigDecimal"];
  depositedToken1: Scalars["BigDecimal"];
  feeGrowthInside0LastX128: Scalars["BigInt"];
  feeGrowthInside1LastX128: Scalars["BigInt"];
  id: Scalars["ID"];
  liquidity: Scalars["BigInt"];
  owner: Scalars["Bytes"];
  pool: Pool;
  position: Position;
  timestamp: Scalars["BigInt"];
  transaction: Transaction;
  withdrawnToken0: Scalars["BigDecimal"];
  withdrawnToken1: Scalars["BigDecimal"];
};
// ...
export type Query = {
  __typename?: "Query";
  /** Access to subgraph metadata */
  _meta?: Maybe<_Meta_>;
  bundle?: Maybe<Bundle>;
  bundles: Array<Bundle>;
  burn?: Maybe<Burn>;
  burns: Array<Burn>;
  collect?: Maybe<Collect>;
  collects: Array<Collect>;
  factories: Array<Factory>;
  factory?: Maybe<Factory>;
  flash?: Maybe<Flash>;
  flashes: Array<Flash>;
  mint?: Maybe<Mint>;
  mints: Array<Mint>;
  pool?: Maybe<Pool>;
  poolDayData?: Maybe<PoolDayData>;
  poolDayDatas: Array<PoolDayData>;
  poolHourData?: Maybe<PoolHourData>;
  poolHourDatas: Array<PoolHourData>;
  pools: Array<Pool>;
  position?: Maybe<Position>;
  positionSnapshot?: Maybe<PositionSnapshot>;
  positionSnapshots: Array<PositionSnapshot>;
  positions: Array<Position>;
  swap?: Maybe<Swap>;
  swaps: Array<Swap>;
  tick?: Maybe<Tick>;
  tickDayData?: Maybe<TickDayData>;
  tickDayDatas: Array<TickDayData>;
  tickHourData?: Maybe<TickHourData>;
  tickHourDatas: Array<TickHourData>;
  ticks: Array<Tick>;
  token?: Maybe<Token>;
  tokenDayData?: Maybe<TokenDayData>;
  tokenDayDatas: Array<TokenDayData>;
  tokenHourData?: Maybe<TokenHourData>;
  tokenHourDatas: Array<TokenHourData>;
  tokens: Array<Token>;
  transaction?: Maybe<Transaction>;
  transactions: Array<Transaction>;
  uniswapDayData?: Maybe<UniswapDayData>;
  uniswapDayDatas: Array<UniswapDayData>;
};
// ...
export type Swap = {
  __typename?: "Swap";
  amount0: Scalars["BigDecimal"];
  amount1: Scalars["BigDecimal"];
  amountUSD: Scalars["BigDecimal"];
  id: Scalars["ID"];
  logIndex?: Maybe<Scalars["BigInt"]>;
  origin: Scalars["Bytes"];
  pool: Pool;
  recipient: Scalars["Bytes"];
  sender: Scalars["Bytes"];
  sqrtPriceX96: Scalars["BigInt"];
  tick: Scalars["BigInt"];
  timestamp: Scalars["BigInt"];
  token0: Token;
  token1: Token;
  transaction: Transaction;
};
// ...
export type Tick = {
  __typename?: "Tick";
  collectedFeesToken0: Scalars["BigDecimal"];
  collectedFeesToken1: Scalars["BigDecimal"];
  collectedFeesUSD: Scalars["BigDecimal"];
  createdAtBlockNumber: Scalars["BigInt"];
  createdAtTimestamp: Scalars["BigInt"];
  feeGrowthOutside0X128: Scalars["BigInt"];
  feeGrowthOutside1X128: Scalars["BigInt"];
  feesUSD: Scalars["BigDecimal"];
  id: Scalars["ID"];
  liquidityGross: Scalars["BigInt"];
  liquidityNet: Scalars["BigInt"];
  liquidityProviderCount: Scalars["BigInt"];
  pool: Pool;
  poolAddress?: Maybe<Scalars["String"]>;
  price0: Scalars["BigDecimal"];
  price1: Scalars["BigDecimal"];
  tickIdx: Scalars["BigInt"];
  untrackedVolumeUSD: Scalars["BigDecimal"];
  volumeToken0: Scalars["BigDecimal"];
  volumeToken1: Scalars["BigDecimal"];
  volumeUSD: Scalars["BigDecimal"];
};

export type TickDayData = {
  __typename?: "TickDayData";
  date: Scalars["Int"];
  feeGrowthOutside0X128: Scalars["BigInt"];
  feeGrowthOutside1X128: Scalars["BigInt"];
  feesUSD: Scalars["BigDecimal"];
  id: Scalars["ID"];
  liquidityGross: Scalars["BigInt"];
  liquidityNet: Scalars["BigInt"];
  pool: Pool;
  tick: Tick;
  volumeToken0: Scalars["BigDecimal"];
  volumeToken1: Scalars["BigDecimal"];
  volumeUSD: Scalars["BigDecimal"];
};
// ...
export type TickHourData = {
  __typename?: "TickHourData";
  feesUSD: Scalars["BigDecimal"];
  id: Scalars["ID"];
  liquidityGross: Scalars["BigInt"];
  liquidityNet: Scalars["BigInt"];
  periodStartUnix: Scalars["Int"];
  pool: Pool;
  tick: Tick;
  volumeToken0: Scalars["BigDecimal"];
  volumeToken1: Scalars["BigDecimal"];
  volumeUSD: Scalars["BigDecimal"];
};
// ...
export type Token = {
  __typename?: "Token";
  decimals: Scalars["BigInt"];
  derivedETH: Scalars["BigDecimal"];
  feesUSD: Scalars["BigDecimal"];
  id: Scalars["ID"];
  name: Scalars["String"];
  poolCount: Scalars["BigInt"];
  symbol: Scalars["String"];
  tokenDayData: Array<TokenDayData>;
  totalSupply: Scalars["BigInt"];
  totalValueLocked: Scalars["BigDecimal"];
  totalValueLockedUSD: Scalars["BigDecimal"];
  totalValueLockedUSDUntracked: Scalars["BigDecimal"];
  txCount: Scalars["BigInt"];
  untrackedVolumeUSD: Scalars["BigDecimal"];
  volume: Scalars["BigDecimal"];
  volumeUSD: Scalars["BigDecimal"];
  whitelistPools: Array<Pool>;
};
// ...
export type _Block_ = {
  __typename?: "_Block_";
  /** The hash of the block */
  hash?: Maybe<Scalars["Bytes"]>;
  /** The block number */
  number: Scalars["Int"];
  /** Integer representation of the timestamp stored in blocks for the chain */
  timestamp?: Maybe<Scalars["Int"]>;
};

/** The type for the top-level _meta field */
export type _Meta_ = {
  __typename?: "_Meta_";
  /**
   * Information about a specific subgraph block. The hash of the block
   * will be null if the _meta field has a block constraint that asks for
   * a block number. It will be filled if the _meta field has no block constraint
   * and therefore asks for the latest  block
   */
  block: _Block_;
  /** The deployment ID */
  deployment: Scalars["String"];
  /** If `true`, the subgraph encountered indexing errors at some past block */
  hasIndexingErrors: Scalars["Boolean"];
};
// ...
export type FeeTierDistributionQuery = {
  __typename?: "Query";
  _meta?: {
    __typename?: "_Meta_";
    block: { __typename?: "_Block_"; number: number };
  };
  asToken0: Array<{
    __typename?: "Pool";
    feeTier: any;
    totalValueLockedToken0: any;
    totalValueLockedToken1: any;
  }>;
  asToken1: Array<{
    __typename?: "Pool";
    feeTier: any;
    totalValueLockedToken0: any;
    totalValueLockedToken1: any;
  }>;
};